--- src/main/java/com/kh/codelit/community/adminBoard/model/vo/Notice.java
@@ -1,22 +1,0 @@
-package com.kh.codelit.community.adminBoard.model.vo;
-
-import java.util.Date;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Notice {
-	
-	private int noticeNo;
-	private String refMemberId;
-	private String noticeTitle;
-	private String noticeContent;
-	private Date noticeDate;
-	private int noticeCount;
-	private int rownum;
-	
-}
