+++ src/main/java/com/kh/codelit/community/notice/model/dao/AdminBoardDAO.java
@@ -1,0 +1,22 @@
+package com.kh.codelit.community.notice.model.dao;
+
+import java.util.List;
+import java.util.Map;
+
+import com.kh.codelit.community.notice.model.vo.Notice;
+
+public interface AdminBoardDAO {
+
+	int insertBoard(Notice notice);
+
+	int getListCount();
+
+	List<Notice> noticeList(Map<String, Object> param);
+
+	Notice selectOneNotice(int noticeNo);
+
+	int delete(int noticeNo);
+
+	int update(Notice notice);
+
+}
