+++ src/main/java/com/kh/codelit/community/student/controller/StudentBoardController.java
@@ -1,0 +1,18 @@
+package com.kh.codelit.community.student.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import com.kh.codelit.community.notice.model.service.AdminBoardService;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Controller
+@RequestMapping("/community")
+@Slf4j
+public class StudentBoardController {
+	
+	
+	
+}
