+++ src/main/java/com/kh/codelit/community/student/model/dao/CommentDAO.java
@@ -1,0 +1,5 @@
+package com.kh.codelit.community.student.model.dao;
+
+public interface CommentDAO {
+
+}
