+++ src/main/java/com/kh/codelit/community/student/model/dao/StudentBoardDAO.java
@@ -1,0 +1,5 @@
+package com.kh.codelit.community.student.model.dao;
+
+public interface StudentBoardDAO {
+
+}
