+++ src/main/java/com/kh/codelit/community/student/model/dao/StudentBoardDAOImpl.java
@@ -1,0 +1,8 @@
+package com.kh.codelit.community.student.model.dao;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class StudentBoardDAOImpl implements StudentBoardDAO {
+
+}
