+++ src/main/java/com/kh/codelit/community/student/model/service/CommentService.java
@@ -1,0 +1,5 @@
+package com.kh.codelit.community.student.model.service;
+
+public interface CommentService {
+
+}
