+++ src/main/java/com/kh/codelit/community/student/model/service/CommentServiceImpl.java
@@ -1,0 +1,15 @@
+package com.kh.codelit.community.student.model.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.kh.codelit.community.student.model.dao.CommentDAO;
+
+@Service
+public class CommentServiceImpl implements CommentService {
+
+	@Autowired
+	private CommentDAO dao;
+
+	
+}
