+++ src/main/java/com/kh/codelit/community/student/model/service/StudentBoardServiceImpl.java
@@ -1,0 +1,10 @@
+package com.kh.codelit.community.student.model.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class StudentBoardServiceImpl implements StudentBoardService {
+	
+	
+	
+}
