+++ src/main/java/com/kh/codelit/community/student/model/vo/StudentBoard.java
@@ -1,0 +1,24 @@
+package com.kh.codelit.community.student.model.vo;
+
+import java.util.Date;
+
+import com.kh.codelit.community.notice.model.vo.Notice;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class StudentBoard {
+
+	private int stdBrdNo;
+	private String refMemberId;
+	private int refLectureNo;
+	private String stdBrdTitle;
+	private String stdBrdContent;
+	private Date stdBrdDate;
+	private int stdBrdCount;
+	
+}
