--- src/main/java/com/kh/codelit/community/studentBoard/controller/StudentBoardController.java
@@ -1,18 +1,0 @@
-package com.kh.codelit.community.studentBoard.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import com.kh.codelit.community.adminBoard.model.service.AdminBoardService;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Controller
-@RequestMapping("/community")
-@Slf4j
-public class StudentBoardController {
-	
-	
-	
-}
