--- src/main/java/com/kh/codelit/community/studentBoard/model/dao/CommentDAO.java
@@ -1,5 +1,0 @@
-package com.kh.codelit.community.studentBoard.model.dao;
-
-public interface CommentDAO {
-
-}
