--- src/main/java/com/kh/codelit/community/studentBoard/model/dao/CommentDAOImpl.java
@@ -1,8 +1,0 @@
-package com.kh.codelit.community.studentBoard.model.dao;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class CommentDAOImpl implements CommentDAO {
-
-}
