--- src/main/java/com/kh/codelit/community/studentBoard/model/dao/StudentBoardDAOImpl.java
@@ -1,8 +1,0 @@
-package com.kh.codelit.community.studentBoard.model.dao;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class StudentBoardDAOImpl implements StudentBoardDAO {
-
-}
