--- src/main/java/com/kh/codelit/community/studentBoard/model/service/CommentService.java
@@ -1,5 +1,0 @@
-package com.kh.codelit.community.studentBoard.model.service;
-
-public interface CommentService {
-
-}
