--- src/main/java/com/kh/codelit/community/studentBoard/model/service/StudentBoardServiceImpl.java
@@ -1,10 +1,0 @@
-package com.kh.codelit.community.studentBoard.model.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class StudentBoardServiceImpl implements StudentBoardService {
-	
-	
-	
-}
